fix VS build warnings in tests

+++ tests/test004.cpp
@@ -34,17 +34,17 @@ int main()
     std::vector<std::string> strs;
 
     ints = doc.GetColumn<int>(0);
-    unittest::ExpectEqual(int, ints.size(), 2);
+    unittest::ExpectEqual(size_t, ints.size(), 2);
     unittest::ExpectEqual(int, ints.at(0), 3);
     unittest::ExpectEqual(int, ints.at(1), 4);
 
     strs = doc.GetColumn<std::string>(1);
-    unittest::ExpectEqual(int, strs.size(), 2);
+    unittest::ExpectEqual(size_t, strs.size(), 2);
     unittest::ExpectEqual(std::string, strs.at(0), "9");
     unittest::ExpectEqual(std::string, strs.at(1), "16");
 
     ints = doc.GetColumn<int>("C");
-    unittest::ExpectEqual(int, ints.size(), 2);
+    unittest::ExpectEqual(size_t, ints.size(), 2);
     unittest::ExpectEqual(int, ints.at(0), 81);
     unittest::ExpectEqual(int, ints.at(1), 256);
