+++ tests/test006.cpp
@@ -33,13 +33,13 @@ int main()
     std::vector<std::string> strs;
 
     ints = doc.GetRow<int>("1");
-    unittest::ExpectEqual(int, ints.size(), 3);
+    unittest::ExpectEqual(size_t, ints.size(), 3);
     unittest::ExpectEqual(int, ints.at(0), 3);
     unittest::ExpectEqual(int, ints.at(1), 9);
     unittest::ExpectEqual(int, ints.at(2), 81);
 
     strs = doc.GetRow<std::string>(1);
-    unittest::ExpectEqual(int, strs.size(), 3);
+    unittest::ExpectEqual(size_t, strs.size(), 3);
     unittest::ExpectEqual(std::string, strs.at(0), "4");
     unittest::ExpectEqual(std::string, strs.at(1), "16");
     unittest::ExpectEqual(std::string, strs.at(2), "256");
